--- station/builtin_call.cpp
@@ -1,142 +1,0 @@
-/**
- * @file builtin_call.cpp
- * @brief The built in call objects for every client
- * 
- * This includes the 'set' and 'respond' commands.
- * 
- * @license{This project is released under the MIT License.}
- */
-
-
-#define RM_NO_WX
-
-
-#include "rm/builtin_call.hpp"
-
-
-/**
- * @brief Constructs with a client
- * 
- * @param cli The client
- */
-rmBuiltinCallSet::rmBuiltinCallSet(rmClient* cli)
-                 :rmCall("set", nullptr)
-{
-    client = cli;
-}
-
-/**
- * @brief Invokes the callback of the object
- * 
- * @param argc Argument count
- * @param argv Tokens
- */
-void rmBuiltinCallSet::invoke(int argc, char* argv[]) {
-    if(argc != 2)
-        return;
-    
-    rmAttribute* attr = client->getAttribute(argv[0]);
-    if(attr != nullptr) {
-        rmAttributeData prev = attr->getValue();
-        attr->setValue(argv[1]);
-        if(attr->getValue().i != prev.i) {
-            rmAttributeNotifier* noti = attr->getNotifier();
-            if(noti != nullptr)
-                noti->triggerCallback();
-        }
-    }
-}
-
-
-
-
-static void echo(rmClient *client, int status, int argc, char *argv[]) {
-    char str[128];
-    int i=0, j=0, k=0;
-    while(i < argc) {
-        char c;
-        while((c = argv[i][j++]) != '\0')
-            str[k++] = c;
-        if(i < argc - 1)
-            str[k++] = ' ';
-        i++;
-        j = 0;
-        if(k == 127)
-            break;
-    }
-    str[k] = '\0';
-    str[127] = '\0';
-    client->echo(str, status);
-}
-
-
-
-
-/**
- * @brief Constructs with a client
- * 
- * @param cli The client
- */
-rmBuiltinCallEcho::rmBuiltinCallEcho(rmClient* cli)
-                  :rmCall("echo", nullptr)
-{
-    client = cli;
-}
-
-/**
- * @brief Invokes the callback of the object
- * 
- * @param argc Argument count
- * @param argv Tokens
- */
-void rmBuiltinCallEcho::invoke(int argc, char* argv[]) {
-    echo(client, 0, argc, argv);
-}
-
-
-
-
-/**
- * @brief Constructs with a client
- * 
- * @param cli The client
- */
-rmBuiltinCallWarn::rmBuiltinCallWarn(rmClient* cli)
-                  :rmCall("warn", nullptr)
-{
-    client = cli;
-}
-
-/**
- * @brief Invokes the callback of the object
- * 
- * @param argc Argument count
- * @param argv Tokens
- */
-void rmBuiltinCallWarn::invoke(int argc, char* argv[]) {
-    echo(client, 2, argc, argv);
-}
-
-
-
-
-/**
- * @brief Constructs with a client
- * 
- * @param cli The client
- */
-rmBuiltinCallError::rmBuiltinCallError(rmClient* cli)
-                  :rmCall("error", nullptr)
-{
-    client = cli;
-}
-
-/**
- * @brief Invokes the callback of the object
- * 
- * @param argc Argument count
- * @param argv Tokens
- */
-void rmBuiltinCallError::invoke(int argc, char* argv[]) {
-    echo(client, 1, argc, argv);
-}
